fix(647): Moved LoadBalancerEditer component to different path
